reorganising and renaming for the client side cloudinary util
